Fix single-net interface net naming (#201)
Before: check that interface has 1 field and that it is a net
After: check that interface has 1 net field (and no interfaces)

This allows for adding metadata fields without changing the net naming.
This should (hopefully) not cause that much churn because we haven't
really used metadata fields in interfaces yet.

Signed-off-by: akhilles <akhilvelagapudi@gmail.com>